Revert "UI updates"
