+++ README.md
@@ -1,6 +1,6 @@
 Notes:
 
-[ ] we need to add a 'NEXT_PUBLIC_SITE_URL' when we have one or else the template file routing won't work
+[ ] we need to add a 'NEXT_PUBLIC_BASE_URL' when we have one or else the template file routing won't work
 [ ] project limit is env var stored in .env.local: NEXT_PUBLIC_PROJECT_LIMIT=2
 [ ] env vars need an anthropic api key
