+++ src/app/services/cloudfront.ts
@@ -8,9 +8,10 @@ import {
   PriceClass,
   SSLSupportMethod,
   MinimumProtocolVersion,
+  CreateCloudFrontOriginAccessIdentityCommand,
 } from "@aws-sdk/client-cloudfront";
 import { ACMClient, ListCertificatesCommand } from "@aws-sdk/client-acm";
-// import { S3Client } from "@aws-sdk/client-s3";
+import { S3Client, PutBucketPolicyCommand } from "@aws-sdk/client-s3";
 
 const region = process.env.CUSTOM_REGION;
 
@@ -50,47 +51,52 @@ export class CloudFrontService {
         domainNames,
       });
 
-      // Use static OAI from environment variables
-      const oaiId = process.env.CLOUDFRONT_ORIGIN_ACCESS_ID;
+      // Setup OAI and bucket policy
+      console.log("[CloudFront] Setting up Origin Access Identity");
+      const oaiResponse = await this.setupOriginAccessIdentity(
+        "default-oai",
+        "system"
+      );
+      const oaiId = oaiResponse.CloudFrontOriginAccessIdentity?.Id;
+      const oaiS3CanonicalUserId =
+        oaiResponse.CloudFrontOriginAccessIdentity?.S3CanonicalUserId;
 
-      if (!oaiId) {
-        throw new Error(
-          "CloudFront OAI configuration is missing in environment variables"
-        );
+      if (!oaiId || !oaiS3CanonicalUserId) {
+        throw new Error("Failed to create Origin Access Identity");
       }
 
-      console.log("[CloudFront] Using OAI ID:", oaiId);
+      console.log("[CloudFront] Got OAI ID:", oaiId);
 
       // Update S3 bucket policy to allow OAI access to all site paths
-      // const s3Client = new S3Client({
-      //   region: process.env.CUSTOM_REGION,
-      //   credentials: {
-      //     accessKeyId: process.env.CUSTOM_ACCESS_KEY_ID ?? "",
-      //     secretAccessKey: process.env.CUSTOM_SECRET_ACCESS_KEY ?? "",
-      //   },
-      // });
+      const s3Client = new S3Client({
+        region: process.env.CUSTOM_REGION,
+        credentials: {
+          accessKeyId: process.env.CUSTOM_ACCESS_KEY_ID ?? "",
+          secretAccessKey: process.env.CUSTOM_SECRET_ACCESS_KEY ?? "",
+        },
+      });
 
-      // const bucketPolicy = {
-      //   Version: "2012-10-17",
-      //   Statement: [
-      //     {
-      //       Sid: "AllowCloudFrontAccess",
-      //       Effect: "Allow",
-      //       Principal: {
-      //         AWS: `arn:aws:iam::cloudfront:user/CloudFront Origin Access Identity ${process.env.CLOUDFRONT_ORIGIN_ACCESS_ID}`,
-      //       },
-      //       Action: "s3:GetObject",
-      //       Resource: `arn:aws:s3:::${process.env.S3_BUCKET_NAME}/*`,
-      //     },
-      //   ],
-      // };
+      const bucketPolicy = {
+        Version: "2012-10-17",
+        Statement: [
+          {
+            Sid: "AllowCloudFrontAccess",
+            Effect: "Allow",
+            Principal: {
+              CanonicalUser: oaiS3CanonicalUserId,
+            },
+            Action: "s3:GetObject",
+            Resource: `arn:aws:s3:::${process.env.S3_BUCKET_NAME}/*`,
+          },
+        ],
+      };
 
-      // await s3Client.send(
-      //   new PutBucketPolicyCommand({
-      //     Bucket: process.env.S3_BUCKET_NAME,
-      //     Policy: JSON.stringify(bucketPolicy),
-      //   })
-      // );
+      await s3Client.send(
+        new PutBucketPolicyCommand({
+          Bucket: process.env.S3_BUCKET_NAME,
+          Policy: JSON.stringify(bucketPolicy),
+        })
+      );
 
       // Determine the domain names to use
       let aliasItems: string[] = [];
@@ -180,7 +186,6 @@ export class CloudFrontService {
       throw error;
     }
   }
-
   /**
    * Gets the status of a CloudFront distribution
    */
@@ -213,6 +218,32 @@ export class CloudFrontService {
       return response.Distribution;
     } catch (error) {
       console.error("Error updating distribution:", error);
+      throw error;
+    }
+  }
+
+  private static async setupOriginAccessIdentity(
+    domainName: string,
+    userId: string
+  ) {
+    try {
+      // Create OAI
+      const oaiCommand = new CreateCloudFrontOriginAccessIdentityCommand({
+        CloudFrontOriginAccessIdentityConfig: {
+          CallerReference: `${userId}-${Date.now()}-oai`,
+          Comment: `OAI for ${domainName}`,
+        },
+      });
+      const oaiResponse = await cloudfrontClient.send(oaiCommand);
+      const oaiId = oaiResponse.CloudFrontOriginAccessIdentity?.Id;
+
+      if (!oaiId) {
+        throw new Error("Failed to create Origin Access Identity");
+      }
+
+      return oaiResponse;
+    } catch (error) {
+      console.error("Error setting up Origin Access Identity:", error);
       throw error;
     }
   }
