--- src/components/breadcrumbs.tsx
@@ -1,45 +1,0 @@
-import * as React from "react"
-import Link from "next/link"
-import { ChevronRight } from "lucide-react"
-import { cn } from "@/lib/utils"
-
-export interface BreadcrumbItem {
-  label: string
-  href?: string
-  isActive?: boolean
-}
-
-interface BreadcrumbsProps {
-  items: BreadcrumbItem[]
-  className?: string
-}
-
-export function Breadcrumbs({ items, className }: BreadcrumbsProps) {
-  return (
-    <nav className={cn("flex items-center space-x-2 text-sm text-muted-foreground", className)}>
-      {items.map((item, index) => (
-        <React.Fragment key={item.label}>
-          {index > 0 && <ChevronRight className="h-4 w-4" />}
-          {item.href ? (
-            <Link
-              href={item.href}
-              className={cn(
-                "flex items-center hover:text-foreground",
-                item.isActive && "text-foreground font-bold"
-              )}
-            >
-              {item.label}
-            </Link>
-          ) : (
-            <span className={cn(
-              "text-foreground",
-              item.isActive && "font-bold"
-            )}>
-              {item.label}
-            </span>
-          )}
-        </React.Fragment>
-      ))}
-    </nav>
-  )
-}
