+++ src/hooks/useWebsites.ts
@@ -42,7 +42,8 @@ export function useWebsites(userId: string) {
           )
         `
         )
-        .eq("user_id", userId);
+        .eq("user_id", userId)
+        .limit(1);
       return data as Website[];
     },
     staleTime: 1000 * 60 * 5,
